removed boolean magic

+++ src/Day05.kt
@@ -19,7 +19,7 @@ fun main() = day(5) {
 		val order = orderLines.map { it.split("|").mapToInts() }.groupBy({ it[0] }, { it[1] })
 		val books = bookLines.map { it.split(",").mapToInts() }
 		fun Int.mustBeBefore(other: Int): Boolean = order[this]?.contains(other) ?: false
-		val comparator = Comparator<Int> { l, r -> r.mustBeBefore(l) - l.mustBeBefore(r) }
+		val comparator = Comparator<Int> { l, r -> r.mustBeBefore(l).toInt() - l.mustBeBefore(r).toInt() }
 
 		books.sumOf { pages ->
 			val ordered = pages.windowed(2) { (a, b) -> a.mustBeBefore(b) }.all { it }
